Add support for queuing entire spotify playlists (#186)
This adds support for queuing entire spotify playlists for playback with !playlist. Each song will be searched for with youtube-dl and then queued for playback.

Requires spotify app credentials to use their API.

Should be extensible to other types of playlists by adding more regexps that map to other services.

+++ plugins/MusicPlayer/player.js
@@ -1,6 +1,8 @@
 const YoutubeDL = require('youtube-dl');
 const Discord = require('discord.js');
 const MemoryStream = require('memorystream');
+const SpotifyWebApi = require('spotify-web-api-node');
+const AuthDetails = require("../../auth.js").getAuthDetails();
 
 let options = false;
 const MUSIC_CHANNEL_NAME = (options && options.musicChannelName) || 'music';
@@ -16,7 +18,9 @@ exports.commands = [
     "queue",
     "dequeue",
     "pause",
-    "resume"
+    "resume",
+    "playlist",
+    "shuffle"
 ]
 
 function getResultTitle(result){
@@ -65,6 +69,13 @@ function getUserVoiceChannel(msg) {
  */
 function wrap(text) {
 	return '```\n' + text.replace(/`/g, '`' + String.fromCharCode(8203)) + '\n```';
+}
+
+function shuffleArray(array) {
+    for (let i = array.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [array[i], array[j]] = [array[j], array[i]];
+    }
 }
 
 class Player {
@@ -136,10 +147,18 @@ class Player {
             console.log("YoutubeDL Stream end");
         })
         setTimeout(()=>{
-            if(this.queue[0].response.channel){
-                const embed = generateResultEmbed('Now Playing',video_info,this.queue[0].queuer);
-                this.queue[0].response.channel.send('',embed);
-                this.queue[0].response.delete();
+            if(this.queue[0].response){
+                if(this.queue[0].response.channel){
+                    const embed = generateResultEmbed('Now Playing',video_info,this.queue[0].queuer);
+                    this.queue[0].response.channel.send('',embed);
+                    this.queue[0].response.delete();
+                }
+            } else {
+                let msg_channel = this.voiceChannel.guild.channels.cache.find((v)=>v.type == "text" && v.name === MUSIC_CHANNEL_NAME);
+                if(msg_channel){
+                    const embed = generateResultEmbed('Now Playing',video_info,this.queue[0].queuer);
+                    msg_channel.send('',embed);
+                }
             }
         const dispatcher = connection.play(buffer,{bitrate:'auto',volume:true});
         this.stream = stream;
@@ -262,19 +281,40 @@ exports.skip = {
 }
 
 exports.queue = {
-    description: "prints the current music queue for this server",
-	process: function(client, msg, suffix) {
+    usage: "[-s]",
+    description: "prints the current music queue for this server. -s to print in concise form",
+	process: async function(client, msg, suffix) {
         let player = getPlayer(msg.guild.id);
         const length = player.queue.length;
         var count = 0;
-        let msg_maker = (msg)=>{
-            if(count == length) return;
-            const song = player.queue[count];
-            const title = count == 0 ? 'Now Playing:' : `${count}:`;
-            count += 1;
-            msg.channel.send('',generateResultEmbed(title,song.info,song.queuer)).then(msg_maker);
+        if(suffix.includes("-s")){
+            let list = player.queue.map((song)=>{
+                let entry = (count == 0 ? 'Now Playing: ' : `${count}: `) + getResultTitle(song.info) + '\n';
+                count++;
+                return entry;
+            });
+            let response = "";
+            while(list.length > 0) {
+                let entry = list.shift();
+                let newmsg = response + entry;
+                if(newmsg.length > (1024 - 8)){
+                    await msg.channel.send(response);
+                    response = entry;
+                } else {
+                    response = newmsg;
+                }
+            }
+            msg.channel.send(response);
+        } else {
+            let msg_maker = (msg)=>{
+                if(count == length) return;
+                const song = player.queue[count];
+                const title = count == 0 ? 'Now Playing:' : `${count}:`;
+                count += 1;
+                msg.channel.send('',generateResultEmbed(title,song.info,song.queuer)).then(msg_maker);
+            }
+            msg_maker(msg);
         }
-        msg_maker(msg);
     }
 }
 
@@ -340,5 +380,134 @@ exports.resume = {
         // Resume.
         msg.channel.send( wrap('Playback resumed.'));
         player.resume();
+    }
+}
+
+exports.playlist = {
+    usage: "`<https://open.spotify.com/playlist/...|spotify:playlist:...>`",
+    description: "plays spotify playlists",
+    process: async function(client, msg, suffix, isEdit) {
+        if(isEdit) return;
+        // Make sure the suffix exists.
+        if (!suffix) return msg.channel.send( wrap('No playlist specified!'));
+
+        // Make sure the user is in voice.
+        var arr = msg.guild.channels.cache.filter((v)=>v.type == "voice").filter((v)=>v.members.has(msg.author.id));
+        let responseChannel = msg.guild.channels.cache.find((v)=>v.type == "text" && v.name === MUSIC_CHANNEL_NAME) || msg.channel;
+        if (arr.length == 0) return msg.channel.send( wrap('You\'re not in a voice channel.'));
+
+        var show_playlist = true;
+
+        var playlist_id = null;
+        let uri_re = /spotify:playlist:(\w+)/i;
+        let uri = uri_re.exec(suffix);
+        if(uri) {
+            playlist_id = uri[1];
+        } else {
+            let link_re = /https:\/\/open.spotify.com\/playlist\/(\w+).*/i
+            let link = link_re.exec(suffix);
+            if(link) {
+                playlist_id = link[1];
+            }
+            //Don't show our playlist embed with an open spotify link since discord makes a nice embed.
+            show_playlist = false;
+        }
+        if(!playlist_id){
+            return msg.channel.send("This doesn't look like a spotify playlist to me...");
+        } else {
+            try {
+                var spotifyApi = new SpotifyWebApi({
+                    clientId: AuthDetails.spotify_client_id,
+                    clientSecret: AuthDetails.spotify_client_secret
+                });
+                console.log("requesting spotify access token");
+                let result = await spotifyApi.clientCredentialsGrant();
+                spotifyApi.setAccessToken(result.body['access_token']);
+                console.log("requesting playlist details for " + playlist_id);
+                let playlist = await spotifyApi.getPlaylist(playlist_id);
+                var embed = null;
+
+                if(show_playlist){
+                    embed = {
+                        embed: {
+                            color: 0x1db954,
+                            author: {
+                                name: playlist.body.owner.display_name,
+                            },
+                            url: playlist.body.external_urls.spotify,
+                            image: {
+                                url: playlist.body.images[0].url,
+                            },
+                            title: playlist.body.name
+                        }
+                    };
+                }
+                let response = await msg.channel.send("Processing playlist: 0 of " + playlist.body.tracks.items.length,embed);
+                
+                // Now queue up the songs for playback
+                let player = getPlayer(msg.guild.id);
+                console.log("Starting search of youtube for tracks in this playlist...");
+                let songs = []
+                //Kick off the searches in parallel.
+                let searches = new Promise(async (resolve) => {
+                    for(item of playlist.body.tracks.items) {
+                        let song_search = 'ytsearch1:' + item.track.name + ' ' + item.track.artists[0].name;
+                        console.log("Searching for " + item.track.name + ' ' + item.track.artists[0].name);
+                        let promise = new Promise((resolve, reject) => {
+                            YoutubeDL.getInfo(song_search, ['-i','--max-downloads', '1', '--no-playlist', '--no-check-certificate'], (err,info) =>{
+                                if(err){
+                                    reject(err);
+                                } else {
+                                    resolve(info);
+                                }
+                            });
+                        });
+                        songs.push(promise);
+                        //Throttle our queries somewhat so we don't piss off youtube.
+                        await new Promise(r => setTimeout(r, 1000));
+                    }
+                    console.log("Done searching for " + playlist_id);
+                    resolve();
+                });
+                let loaded = 0;
+                let edited = true;
+                for(song of songs) {
+                    try {
+                        let info = await song;
+                        player.enqueue(client,msg,null,info);
+                    } catch (e) {
+                        console.log("failed trying to search for a song!");
+                        console.error(e);
+                    }
+                    loaded++;
+                    if(edited){
+                        edited = false;
+                        response.edit("Processing playlist: " + loaded + " of " + playlist.body.tracks.items.length,embed).then(()=>{edited = true;});
+                    }
+                }
+                console.log("All tracks in playlist " + playlist_id + " queued");
+                response.edit("playlist queued",embed);
+            } catch(e) {
+                if(!AuthDetails.spotify_client_id || !AuthDetails.spotify_client_secret){
+                    console.log("Missing spotify api credentials. Did you add them to auth.json?");
+                }
+                console.error(e);
+                return msg.channel.send("Couldn't read the playlist :(");
+            }
+        }
+        
+    }
+}
+
+exports.shuffle = {
+    description: "Shuffles the play queue",
+    process: function (client, msg, suffix) {
+        let player = getPlayer(msg.guild.id);
+        if(player) {
+            shuffleArray(player.queue);
+            msg.channel.send("Shuffled the play queue!");
+        } else {
+            msg.channel.send("Couldn't find a player. Are you playing music?");
+        }
     }
 }
